+++ client/app/_helpers/http-interceptor.ts
@@ -1,0 +1,44 @@
+﻿import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+import { HttpInterceptor, HttpHandler, HttpRequest } from '@angular/common/http';
+import { appConfig } from '../app.config';
+import { Observable } from 'rxjs/Observable';
+import 'rxjs/add/operator/catch';
+import 'rxjs/add/operator/map';
+import 'rxjs/add/observable/throw';
+
+const TOKEN_HEADER_KEY = 'Authorization';
+
+@Injectable()
+export class AuthInterceptor implements HttpInterceptor {
+  constructor() { }
+
+  private getToken(): any {
+    let currentUser = JSON.parse(localStorage.getItem('currentUser'));
+    if (currentUser && currentUser.token) {
+     return currentUser.token;
+    }
+    return '';
+  }
+
+  private handleError(error: any) {
+    if (error.status === 401) {
+      // 401 unauthorized response so log user out of client
+      window.location.href = '/login';
+    }
+    return Observable.throw(error._body);
+  }
+
+  intercept(req: HttpRequest<any>, next: HttpHandler) {
+    let authReq = req;
+    console.log(req);
+    const token = this.getToken();
+    authReq = req.clone({ url: appConfig.apiUrl + req.url, headers: req.headers.set(TOKEN_HEADER_KEY, 'Bearer ' + token) });
+    console.log(authReq);
+    return next.handle(authReq);
+  }
+}
+
+export const authInterceptorProviders = [
+  { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true }
+];
