+++ client/app/_helpers/index.ts
@@ -1,3 +1,3 @@
-﻿export * from './custom-http';
+﻿export * from './http-interceptor';
 export * from './validate-password';
 export * from './validate-all-fields';
